adding tools

+++ bin/common.sh
@@ -12,7 +12,7 @@
 THIS=`basename $0`
 THISDIR=`dirname $0`; THISDIR=`cd $THISDIR;pwd`
 BASEDIR="$THISDIR/.."
-JQ_EXE="$THISDIR/tools/jq"
+JQ_EXE="jq"
 
 ##############################################################################
 ##
@@ -148,10 +148,18 @@ stage_api() {
 				echo aliasId: $aliasId
 
 				if [ -f "$STAGING_DIR/aliases.json" ] ; then
-					newAlias=$($JQ_EXE -c --arg id "$aliasId" '.[] | select( .id==$id ) | .'$environment' + { id : $id }' $STAGING_DIR/aliases.json)
-					echo $newAlias > $file.new
-					cat $file $file.new | $JQ_EXE -s add > $file
-					rm -f $file.new
+					## first, let's check if alias exists...if not, let's fail??
+					newAliasEnv=$($JQ_EXE -c --arg id "$aliasId" '.[] | select( .id==$id ) | .'$environment $STAGING_DIR/aliases.json)
+					if [ "$newAliasEnv" != "null" ] ; then
+						echo "Executing: $JQ_EXE -c --arg id "$aliasId" '.[] | select( .id==\$id ) | .'$environment' + { id : \$id }' $STAGING_DIR/aliases.json"
+						newAlias=$($JQ_EXE -c --arg id "$aliasId" '.[] | select( .id==$id ) | .'$environment' + { id : $id }' $STAGING_DIR/aliases.json)
+						echo "got new alias=$newAlias"
+						echo $newAlias > $file.new
+						cat $file $file.new | $JQ_EXE -s add > $file
+						rm -f $file.new
+					else
+						echo "No matching alias id $aliasId for environment $environment...Not doing any alias replacement and ignoring!"
+					fi
 				fi
 			done
